refactor: cnv error details in gaxi
