+++ src/firestore/src/lib.rs
@@ -45,14 +45,13 @@ pub(crate) mod google {
             include!("generated/convert/firestore/convert.rs");
         }
     }
-    pub mod rpc {
-        include!("generated/protos/rpc/google.rpc.rs");
-        include!("generated/convert/rpc/convert.rs");
-    }
     pub mod r#type {
         // TODO(#1414) - decide if we want to generate this as its own directory.
         include!("generated/protos/firestore/google.r#type.rs");
         include!("generated/convert/type/convert.rs");
+    }
+    pub mod rpc {
+        include!("generated/protos/firestore/google.rpc.rs");
     }
 }
