+++ src/storage/src/lib.rs
@@ -78,8 +78,7 @@ pub(crate) mod google {
         include!("generated/convert/type/convert.rs");
     }
     pub mod rpc {
-        include!("control/generated/protos/rpc/google.rpc.rs");
-        include!("control/generated/convert/rpc/convert.rs");
+        include!("generated/protos/storage/google.rpc.rs");
     }
     pub mod storage {
         #[allow(deprecated)]
